处理收益法报告

+++ assess-app/src/main/java/com/copower/pmcc/assess/dal/basis/mapper/SchemeReimbursementMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface SchemeReimbursementMapper {
-    int countByExample(SchemeReimbursementExample example);
+    long countByExample(SchemeReimbursementExample example);
 
     int deleteByExample(SchemeReimbursementExample example);
